Testsuite PoC - modules refactoring
