--- test-poc/base/src/test/java/org/keycloak/test/base/NoAdminUserKeycloakTestServerConfig.java
@@ -1,19 +1,0 @@
-package org.keycloak.test.base;
-
-import org.keycloak.test.framework.server.KeycloakTestServerConfig;
-
-import java.util.Optional;
-
-public class NoAdminUserKeycloakTestServerConfig implements KeycloakTestServerConfig {
-
-    @Override
-    public String adminUserName() {
-        return null;
-    }
-
-    @Override
-    public String adminUserPassword() {
-        return null;
-    }
-
-}
