+++ test-poc/examples/src/test/java/org/keycloak/test/examples/NoAdminUserKeycloakTestServerConfig.java
@@ -1,0 +1,19 @@
+package org.keycloak.test.examples;
+
+import org.keycloak.test.framework.server.KeycloakTestServerConfig;
+
+import java.util.Optional;
+
+public class NoAdminUserKeycloakTestServerConfig implements KeycloakTestServerConfig {
+
+    @Override
+    public String adminUserName() {
+        return null;
+    }
+
+    @Override
+    public String adminUserPassword() {
+        return null;
+    }
+
+}
