+++ test-poc/examples/src/test/java/org/keycloak/test/examples/PagesTest.java
@@ -1,0 +1,25 @@
+package org.keycloak.test.examples;
+
+import org.junit.jupiter.api.Test;
+import org.keycloak.test.framework.annotations.KeycloakIntegrationTest;
+import org.keycloak.test.framework.page.LoginPage;
+import org.keycloak.test.framework.annotations.TestPage;
+import org.keycloak.test.framework.page.WelcomePage;
+
+@KeycloakIntegrationTest
+public class PagesTest {
+
+    @TestPage
+    WelcomePage welcomePage;
+
+    @TestPage
+    LoginPage loginPage;
+
+    @Test
+    public void testLoginFromWelcome() {
+        welcomePage.navigateTo();
+        loginPage.fillLogin("admin", "admin");
+        loginPage.submit();
+    }
+
+}
