+++ test-poc/examples/src/test/resources/container-license-acceptance.txt
