+++ test-poc/examples/src/test/resources/logging.properties
