--- test-poc/framework/src/main/java/org/keycloak/test/framework/TestAdminClient.java
@@ -1,13 +1,0 @@
-package org.keycloak.test.framework;
-
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface TestAdminClient {
-
-}
