+++ test-poc/framework/src/main/java/org/keycloak/test/framework/annotations/KeycloakIntegrationTest.java
@@ -1,0 +1,17 @@
+package org.keycloak.test.framework.annotations;
+
+import org.keycloak.test.framework.server.DefaultKeycloakTestServerConfig;
+import org.keycloak.test.framework.server.KeycloakTestServerConfig;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface KeycloakIntegrationTest {
+
+    Class<? extends KeycloakTestServerConfig> config() default DefaultKeycloakTestServerConfig.class;
+
+}
