+++ test-poc/framework/src/main/java/org/keycloak/test/framework/annotations/KeycloakTestDatabase.java
@@ -1,0 +1,19 @@
+package org.keycloak.test.framework.annotations;
+
+import org.keycloak.test.framework.database.DatabaseConfig;
+import org.keycloak.test.framework.injection.LifeCycle;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface KeycloakTestDatabase {
+
+    Class<? extends DatabaseConfig> config() default DatabaseConfig.class;
+
+    LifeCycle lifecycle() default LifeCycle.GLOBAL;
+
+}
