+++ test-poc/framework/src/main/java/org/keycloak/test/framework/annotations/TestRealm.java
@@ -1,0 +1,20 @@
+package org.keycloak.test.framework.annotations;
+
+import org.keycloak.test.framework.injection.LifeCycle;
+import org.keycloak.test.framework.realm.DefaultRealmConfig;
+import org.keycloak.test.framework.realm.RealmConfig;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface TestRealm {
+
+    Class<? extends RealmConfig> config() default DefaultRealmConfig.class;
+
+    LifeCycle lifecycle() default LifeCycle.CLASS;
+
+}
