+++ test-poc/framework/src/main/java/org/keycloak/test/framework/annotations/TestWebDriver.java
@@ -1,0 +1,10 @@
+package org.keycloak.test.framework.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface TestWebDriver { }
