--- test-poc/framework/src/main/java/org/keycloak/test/framework/page/TestPage.java
@@ -1,11 +1,0 @@
-package org.keycloak.test.framework.page;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface TestPage {
-}
