+++ test-poc/framework/src/main/java/org/keycloak/test/framework/realm/ClientSupplier.java
@@ -4,7 +4,7 @@ import jakarta.ws.rs.core.Response;
 import org.keycloak.admin.client.resource.ClientResource;
 import org.keycloak.admin.client.resource.RealmResource;
 import org.keycloak.representations.idm.ClientRepresentation;
-import org.keycloak.test.framework.TestClient;
+import org.keycloak.test.framework.annotations.TestClient;
 import org.keycloak.test.framework.injection.InstanceContext;
 import org.keycloak.test.framework.injection.LifeCycle;
 import org.keycloak.test.framework.injection.RequestedInstance;
