+++ test-poc/framework/src/main/java/org/keycloak/test/framework/server/AbstractKeycloakTestServerSupplier.java
@@ -1,6 +1,6 @@
 package org.keycloak.test.framework.server;
 
-import org.keycloak.test.framework.KeycloakIntegrationTest;
+import org.keycloak.test.framework.annotations.KeycloakIntegrationTest;
 import org.keycloak.test.framework.database.TestDatabase;
 import org.keycloak.test.framework.injection.InstanceContext;
 import org.keycloak.test.framework.injection.LifeCycle;
