--- test-poc/framework/src/main/java/org/keycloak/test/framework/webdriver/TestWebDriver.java
@@ -1,10 +1,0 @@
-package org.keycloak.test.framework.webdriver;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface TestWebDriver { }
